+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -566,7 +566,7 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
     vtemp = AssignIncrement(scratch, nmat);
     // TODO(JCD): use whatever lambdas are passed in
     /*for (int m = 0; m < nmat; m++) {
-      if (lambda[m] != nullptr) Cache[m] = lambda[m];
+      if (!variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
     }*/
   }
 
@@ -784,7 +784,7 @@ class PTESolverFixedT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
     Tnorm = 1.0;
     // TODO(JCD): use whatever lambdas are passed in
     /*for (int m = 0; m < nmat; m++) {
-      if (lambda[m] != nullptr) Cache[m] = lambda[m];
+      if (variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
     }*/
   }
 
@@ -982,7 +982,7 @@ class PTESolverFixedP : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
     Pequil = P;
     // TODO(JCD): use whatever lambdas are passed in
     /*for (int m = 0; m < nmat; m++) {
-      if (lambda[m] != nullptr) Cache[m] = lambda[m];
+      if (variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
     }*/
   }
 
@@ -1207,7 +1207,7 @@ class PTESolverRhoU : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
     utemp = AssignIncrement(scratch, nmat);
     // TODO(JCD): use whatever lambdas are passed in
     /*for (int m = 0; m < nmat; m++) {
-      if (lambda[m] != nullptr) Cache[m] = lambda[m];
+      if (variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
     }*/
   }