+++ singularity-eos/eos/eos_powermg.hpp
@@ -45,26 +45,43 @@ class PowerMG : public EosBase<PowerMG> {
   }
 
   PowerMG GetOnDevice() { return *this; }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const;
-  PORTABLE_INLINE_FUNCTION Real PressureFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const;
-  PORTABLE_INLINE_FUNCTION Real PressureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
+      const Real rho, const Real temp,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real
-  MinInternalEnergyFromDensity(const Real rho, Real *lambda = nullptr) const;
-  PORTABLE_INLINE_FUNCTION Real EntropyFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const;
+  PressureFromDensityTemperature(const Real rho, const Real temp,
+                                 Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION Real PressureFromDensityInternalEnergy(
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION Real MinInternalEnergyFromDensity(
+      const Real rho, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION Real
+  EntropyFromDensityTemperature(const Real rho, const Real temp,
+                                Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real EntropyFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const {
+      const Real rho, const Real temp,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     return _Cv0;
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const {
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     return _Cv0;
   }
   // added for testing AEM Dec 2023
@@ -72,30 +89,45 @@ class PowerMG : public EosBase<PowerMG> {
   PORTABLE_INLINE_FUNCTION Real AllHugInternalEnergyFromDensity(const Real rho) const;
   PORTABLE_INLINE_FUNCTION Real AllHugTemperatureFromDensity(const Real rho) const;
   // Thermal Bulk Modulus added AEM Dec 2022
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real TBulkModulusFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const;
+      const Real rho, const Real temp,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real BulkModulusFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const;
+      const Real rho, const Real temp,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
   // Thermal expansion coefficient added AEM 2022
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real TExpansionCoeffFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const;
+      const Real rho, const Real temp,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temp, Real *lambda = nullptr) const {
+      const Real rho, const Real temp,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     return robust::ratio(_G0 * _rho0, rho);
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const {
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     return robust::ratio(_G0 * _rho0, rho);
   }
-  PORTABLE_INLINE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                        Real &cv, Real &bmod, const unsigned long output,
-                                        Real *lambda = nullptr) const;
-  PORTABLE_INLINE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
+  template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION void
+  FillEos(Real &rho, Real &temp, Real &energy, Real &press, Real &cv, Real &bmod,
+          const unsigned long output,
+          Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+  template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION void
+  ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
+                         Real &bmod, Real &dpde, Real &dvdt,
+                         Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
   // Generic functions provided by the base class. These contain e.g. the vector
   // overloads that use the scalar versions declared here
   SG_ADD_BASE_CLASS_USINGS(PowerMG)
@@ -117,9 +149,10 @@ class PowerMG : public EosBase<PowerMG> {
     printf("\n\n");
   }
   // Density/Energy from P/T not unique, if used will give error
+  template <typename Indexer_t>
   PORTABLE_INLINE_FUNCTION void
-  DensityEnergyFromPressureTemperature(const Real press, const Real temp, Real *lambda,
-                                       Real &rho, Real &sie) const;
+  DensityEnergyFromPressureTemperature(const Real press, const Real temp,
+                                       Indexer_t &&lambda, Real &rho, Real &sie) const;
   inline void Finalize() {}
   static std::string EosType() { return std::string("PowerMG"); }
   static std::string EosPyType() { return EosType(); }
@@ -256,33 +289,37 @@ PORTABLE_INLINE_FUNCTION Real PowerMG::AllHugTemperatureFromDensity(Real rho) co
   return value;
 }
 
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::InternalEnergyFromDensityTemperature(
-    const Real rho, const Real temp, Real *lambda) const {
+    const Real rho, const Real temp, Indexer_t &&lambda) const {
   Real value = AllHugInternalEnergyFromDensity(rho) +
                _Cv0 * (temp - AllHugTemperatureFromDensity(rho));
   return value;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::PressureFromDensityTemperature(
-    const Real rho, const Real temp, Real *lambda) const {
+    const Real rho, const Real temp, Indexer_t &&lambda) const {
   Real value = AllHugPressureFromDensity(rho) +
                _G0 * _rho0 * _Cv0 * (temp - AllHugTemperatureFromDensity(rho));
   return value;
 }
-PORTABLE_INLINE_FUNCTION Real PowerMG::EntropyFromDensityTemperature(const Real rho,
-                                                                     const Real temp,
-                                                                     Real *lambda) const {
+template <typename Indexer_t>
+PORTABLE_INLINE_FUNCTION Real PowerMG::EntropyFromDensityTemperature(
+    const Real rho, const Real temp, Indexer_t &&lambda) const {
   const Real eta = 1.0 - robust::ratio(_rho0, rho);
   Real value = _S0 - _G0 * _Cv0 * eta + _Cv0 * std::log(temp / _T0);
   return value;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::TExpansionCoeffFromDensityTemperature(
-    const Real rho, const Real temp, Real *lambda) const {
+    const Real rho, const Real temp, Indexer_t &&lambda) const {
   Real value =
       robust::ratio(_Cv0 * _rho0 * _G0, TBulkModulusFromDensityTemperature(rho, temp));
   return value;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::TBulkModulusFromDensityTemperature(
-    const Real rho, const Real temp, Real *lambda) const {
+    const Real rho, const Real temp, Indexer_t &&lambda) const {
   return BulkModulusFromDensityTemperature(rho, temp) -
          _G0 * _G0 * _Cv0 * _rho0 * robust::ratio(_rho0, rho) * temp;
 }
@@ -300,8 +337,9 @@ PowerMG::_compBulkModulusFromDensityTemperature(const Real rho, const Real temp)
   value = robust::ratio(_rho0, rho) * value;
   return value;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::BulkModulusFromDensityTemperature(
-    const Real rho, const Real temp, Real *lambda) const {
+    const Real rho, const Real temp, Indexer_t &&lambda) const {
   const Real eta = 1.0 - robust::ratio(_rho0, rho);
   Real value;
   if (eta <= 0.0) {
@@ -330,8 +368,9 @@ PowerMG::_compBulkModulusFromDensityInternalEnergy(const Real rho, const Real si
   value = robust::ratio(_rho0, rho) * value;
   return value;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::BulkModulusFromDensityInternalEnergy(
-    const Real rho, const Real sie, Real *lambda) const {
+    const Real rho, const Real sie, Indexer_t &&lambda) const {
   const Real eta = 1.0 - robust::ratio(_rho0, rho);
   Real value;
   if (eta <= 0.0) {
@@ -349,8 +388,9 @@ PORTABLE_INLINE_FUNCTION Real PowerMG::BulkModulusFromDensityInternalEnergy(
   return value;
 }
 
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::TemperatureFromDensityInternalEnergy(
-    const Real rho, const Real sie, Real *lambda) const {
+    const Real rho, const Real sie, Indexer_t &&lambda) const {
   Real value = (sie - AllHugInternalEnergyFromDensity(rho)) / _Cv0 +
                AllHugTemperatureFromDensity(rho);
   if (value < 0.0) {
@@ -361,19 +401,22 @@ PORTABLE_INLINE_FUNCTION Real PowerMG::TemperatureFromDensityInternalEnergy(
   }
   return value;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::PressureFromDensityInternalEnergy(
-    const Real rho, const Real sie, Real *lambda) const {
+    const Real rho, const Real sie, Indexer_t &&lambda) const {
   Real value = AllHugPressureFromDensity(rho) +
                _rho0 * _G0 * (sie - AllHugInternalEnergyFromDensity(rho));
   return value;
 }
-PORTABLE_INLINE_FUNCTION Real PowerMG::MinInternalEnergyFromDensity(const Real rho,
-                                                                    Real *lambda) const {
+template <typename Indexer_t>
+PORTABLE_INLINE_FUNCTION Real
+PowerMG::MinInternalEnergyFromDensity(const Real rho, Indexer_t &&lambda) const {
   MinInternalEnergyIsNotEnabled("PowerMG");
   return 0.0;
 }
+template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION Real PowerMG::EntropyFromDensityInternalEnergy(
-    const Real rho, const Real sie, Real *lambda) const {
+    const Real rho, const Real sie, Indexer_t &&lambda) const {
   const Real eta = 1.0 - robust::ratio(_rho0, rho);
   Real value = std::log(TemperatureFromDensityInternalEnergy(rho, sie) / _T0);
   value = _S0 - _G0 * _Cv0 * eta + _Cv0 * value;
@@ -386,18 +429,18 @@ PORTABLE_INLINE_FUNCTION Real PowerMG::EntropyFromDensityInternalEnergy(
   return value;
 }
 // AEM: Give error since function is not well defined
-PORTABLE_INLINE_FUNCTION void
-PowerMG::DensityEnergyFromPressureTemperature(const Real press, const Real temp,
-                                              Real *lambda, Real &rho, Real &sie) const {
+template <typename Indexer_t>
+PORTABLE_INLINE_FUNCTION void PowerMG::DensityEnergyFromPressureTemperature(
+    const Real press, const Real temp, Indexer_t &&lambda, Real &rho, Real &sie) const {
   EOS_ERROR("PowerMG::DensityEnergyFromPressureTemperature: "
             "Not implemented.\n");
 }
 // AEM: We should add entropy and Gruneissen parameters here so that it is complete
 // If we add also alpha and BT, those should also be in here.
-PORTABLE_INLINE_FUNCTION void PowerMG::FillEos(Real &rho, Real &temp, Real &sie,
-                                               Real &press, Real &cv, Real &bmod,
-                                               const unsigned long output,
-                                               Real *lambda) const {
+template <typename Indexer_t>
+PORTABLE_INLINE_FUNCTION void
+PowerMG::FillEos(Real &rho, Real &temp, Real &sie, Real &press, Real &cv, Real &bmod,
+                 const unsigned long output, Indexer_t &&lambda) const {
   const unsigned long input = ~output; // everything that is not output is input
   if (thermalqs::density & output) {
     EOS_ERROR("PowerMG FillEos: Density is required input.\n");
@@ -420,10 +463,11 @@ PORTABLE_INLINE_FUNCTION void PowerMG::FillEos(Real &rho, Real &temp, Real &sie,
 }
 
 // TODO(JMM): pre-cache these rather than recomputing them each time
-PORTABLE_INLINE_FUNCTION
-void PowerMG::ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press,
-                                     Real &cv, Real &bmod, Real &dpde, Real &dvdt,
-                                     Real *lambda) const {
+template <typename Indexer_t>
+PORTABLE_INLINE_FUNCTION void
+PowerMG::ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
+                                Real &bmod, Real &dpde, Real &dvdt,
+                                Indexer_t &&lambda) const {
   // AEM: Added all variables I think should be output eventually
   Real tbmod;
   // Real entropy, alpha, Gamma;