+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - [[PR356]](https://github.com/lanl/singularity-eos/pull/356) Update CMake for proper Kokkos linking in Fortran interface
 
 ### Changed (changing behavior/API/variables/...)
+- [[PR363]](https://github.com/lanl/singularity-eos/pull/363) Template lambda values for scalar calls
 
 ### Infrastructure (changes irrelevant to downstream codes)
 - [[PR329]](https://github.com/lanl/singularity-eos/pull/329) Move vinet tests into analytic test suite